requirements.txt
Django==2.2.3
mysqlclient==1.4.2.post1
djangorestframework==3.10.2
django-rest-auth==0.9.5
django-allauth==0.39.1
stripe==2.33.0
markdown==3.1.1